more

+++ backend/setup.sh
@@ -2,11 +2,33 @@
 
 set -e
 
+
+if [[ -z "${RDS_PASSWORD}" ]]; then
+  echo Missing RDS_PASSWORD
+  exit 1
+fi
+
+if [[ -z "${RDS_HOSTNAME}" ]]; then
+  echo Missing RDS_HOSTNAME
+  exit 1
+fi
+
+if [[ -z "${AWS_ACCESS_KEY_ID}" ]]; then
+  echo Missing AWS_ACCESS_KEY_ID
+  exit 1
+fi
+
+if [[ -z "${AWS_SECRET_ACCESS_KEY}" ]]; then
+  echo Missing AWS_SECRET_ACCESS_KEY
+  exit 1
+fi
+
+
 # Dependencies
 
 sudo add-apt-repository ppa:jonathonf/python-3.6 -y
 sudo apt update
-sudo apt install python3.6 make git nginx -y
+sudo apt install linux-headers-$(uname -r) python3.6 make git nginx awscli keepalived -y
 
 sudo apt dist-upgrade -y
 sudo apt autoremove -y
@@ -34,10 +56,14 @@ python manage.py collectstatic
 
 # Systemd
 
+sed -i -e "s/RDS_PASSWORD=/RDS_PASSWORD=${RDS_PASSWORD}/g" /home/ubuntu/infra/backend/env
+sed -i -e "s/RDS_HOSTNAME=/RDS_HOSTNAME=${RDS_HOSTNAME}/g" /home/ubuntu/infra/backend/env
+
 sudo cp /home/ubuntu/infra/backend/gunicorn.service /etc/systemd/system/gunicorn.service
 sudo systemctl daemon-reload
 sudo systemctl restart gunicorn
 sudo systemctl enable gunicorn
+
 
 # Nginx
 
@@ -47,3 +73,17 @@ sudo cp /home/ubuntu/infra/backend/nginx /etc/nginx/sites-available/gymapplife
 sudo ln -sf /etc/nginx/sites-available/gymapplife /etc/nginx/sites-enabled
 rm -rf /etc/nginx/sites-enabled/default
 sudo systemctl restart nginx
+
+
+# AWS
+
+mkdir -p ~/.aws
+echo [default]> ~/.aws/config
+echo region = us-east-1>> ~/.aws/config
+
+echo [default]> ~/.aws/credentials
+echo aws_access_key_id=$AWS_ACCESS_KEY_ID>> ~/.aws/credentials
+echo aws_secret_access_key=$AWS_SECRET_ACCESS_KEY>> ~/.aws/credentials
+
+
+# Keepalived
